+++ crates/fuzzyfold/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this crate will be documented in this file.
 
 ## development
 - preliminary ff-tledit interface
-- changed ff-timecoures output to *.nxy
+- changed ff-timecourse output to *.nxy
 
 ## [0.4.1] - 2026-02-26
 ## Changed